switched to u64s after noticing the grid is 5x7, not 5x6

+++ 2024/day-25/src/lib.rs
@@ -5,7 +5,7 @@ pub mod part1;
 
 #[derive(Debug)]
 pub struct Puzzle {
-    items: Vec<u32>,
+    items: Vec<u64>,
 }
 
 impl std::str::FromStr for Puzzle {
@@ -19,7 +19,7 @@ impl std::str::FromStr for Puzzle {
                 schematic
                     .chars()
                     .filter(char::is_ascii_punctuation)
-                    .map(|c| c as u32)
+                    .map(|c| c as u64)
                     // '#' has a 1 as it's least significant bit, '.' a zero
                     .fold(0, |acc, c| (c & 1) | (acc << 1))
             })
